Implement `withdraw` function for a proposer
Before sending the funds to the proposer, we should check if
the current state has the right condition for the propoer to
withdraw thd funds of the proposal. The check should include
whether the proposal has been approved and whether 24 hours
has been passed after the vote counting finsihed and the
withdrawal has been refused by the owner etc.