Transfer the issued budget to Commons Budget
The Budget used for the common budget is issued to the `IssuedContract` and
the Commons Budget contract can request the budget to be transferred and
then use it. The Commons Budget requests the budget as much as it needs and
it can request as many time as it need to.